rename rest to webapi to avoid conflict with new shared lib
